Remove client side validation
After some discussion, we decided to remove the client side validation
from incoming user input.

The reason for these changes are to allow for server side changes
without the need for tying the client version to specific and strict
policies.

While the changes here had prevented input from returning responses that
we now known would result in 403 errors, the tradeoff for removal is
felt warranted.

This also allows other tokens outside those defined via the iota
to be provided by the caller.