Adds Report API
Report API now supports reporting batches transactions in a single call.

Documents the ReportResponse return type as well as adding an struct
which supports decoding reporting errors.

Extends the Backend interface to support Reporting.

A helper function has been added to convert incoming Report transactions
to correctly formatted query URL. This is necessary because the URL
formatting expected by 3scale is different than that of Auth and AuthRep because of
the batching support.