Restructure API to remove Options from Transaction
This chnage was made becasue it makes more sense to have the Options as
a standalone structure and not pass them as a field within each
Transaction.

This will now allow the user to provide n amount of available Options to
extend the functionality of the API, while keeping the general and
default use cases simple.

Because we no longer need the functional options on the Transaction, the
API has changed from accepting pointers to Transactions to Transaction
structs. The is more robust since it will not cause the library to panic
in cases where nil is passed by the user.