Use 3scale 'code' as Reason in Auth/AuthRep response
This change was made because often times, the 'Reason' returned from
3scale is an empty string. This provides no context to the end user as
to the cause of the error.

The 'Code' field is always populated and returns some brief information
on why the request failed.

Futhermore, the 'Reason', if it is populated often leaks out secrets and
this reduces the risk of such secrets getting entered into logging
systems.