api: Support timestamp as Unix time
Up to this point we had supported passing a timestamp in a transaction
as a string. This was because 3scale accepted and expected only a specific
string format.

This will change going forward and the Report API will now accept Unix
timestamps, so for a client who is implementing the API in Go, we have
decided that it is better to enforce universally the use of Unix
timestamp over strings and hence the type is changed to int64.